move channel, text channel and voice channel attachable listener into channel sub-package
