+++ javacord-api/src/main/java/org/javacord/api/entity/channel/VoiceChannel.java
@@ -2,7 +2,7 @@ package org.javacord.api.entity.channel;
 
 import org.javacord.api.entity.permission.PermissionType;
 import org.javacord.api.entity.user.User;
-import org.javacord.api.listener.VoiceChannelAttachableListenerManager;
+import org.javacord.api.listener.channel.VoiceChannelAttachableListenerManager;
 
 import java.util.NoSuchElementException;
 import java.util.Optional;
