+++ javacord-api/src/main/java/org/javacord/api/listener/message/MessageEditListener.java
@@ -3,7 +3,7 @@ package org.javacord.api.listener.message;
 import org.javacord.api.event.message.MessageEditEvent;
 import org.javacord.api.listener.GloballyAttachableListener;
 import org.javacord.api.listener.ObjectAttachableListener;
-import org.javacord.api.listener.TextChannelAttachableListener;
+import org.javacord.api.listener.channel.TextChannelAttachableListener;
 import org.javacord.api.listener.server.ServerAttachableListener;
 
 /**
