+++ javacord-api/src/main/java/org/javacord/api/listener/message/reaction/ReactionAddListener.java
@@ -3,7 +3,7 @@ package org.javacord.api.listener.message.reaction;
 import org.javacord.api.event.message.reaction.ReactionAddEvent;
 import org.javacord.api.listener.GloballyAttachableListener;
 import org.javacord.api.listener.ObjectAttachableListener;
-import org.javacord.api.listener.TextChannelAttachableListener;
+import org.javacord.api.listener.channel.TextChannelAttachableListener;
 import org.javacord.api.listener.message.MessageAttachableListener;
 import org.javacord.api.listener.server.ServerAttachableListener;
 import org.javacord.api.listener.user.UserAttachableListener;
