+++ javacord-api/src/main/java/org/javacord/api/listener/user/UserStartTypingListener.java
@@ -3,7 +3,7 @@ package org.javacord.api.listener.user;
 import org.javacord.api.event.user.UserStartTypingEvent;
 import org.javacord.api.listener.GloballyAttachableListener;
 import org.javacord.api.listener.ObjectAttachableListener;
-import org.javacord.api.listener.TextChannelAttachableListener;
+import org.javacord.api.listener.channel.TextChannelAttachableListener;
 import org.javacord.api.listener.server.ServerAttachableListener;
 
 /**
